combine server and parser functionality
server and parser now both live in parser the
script can now successfully initially connect
the f04 and receive live information!!!!
now we need to figure out how to stay in contact
with the server, as it closes connection (likely)
because we didn't respond properly. so, we figure
out how to respond properly, then the server will
be sending us live coordinate information as well!
we will have to make slight alterations to this script
so it continues to use the hash it makes as the ids for
objects will remain constant if we keep a session alive